added map/reduce indexes and browsing webtest

+++ RavenDBBenchmark/NorthwindLoadTestPlugin/Model/OrderLine.cs
@@ -4,7 +4,7 @@ namespace Orders
     {
         public string Product { get; set; }
         public string ProductName { get; set; }
-        public decimal PricePerUnit { get; set; }
+        public double PricePerUnit { get; set; }
         public int Quantity { get; set; }
         public decimal Discount { get; set; }
     }
